test(radio): add tests
