+++ core/src/components/radio-group/test/form/index.html
@@ -35,7 +35,6 @@
               </ion-item>
 
               <ion-item>
-
                 <ion-radio value="griff">Griff</ion-radio>
               </ion-item>
