+++ core/src/components/radio-group/test/search/radio-group.e2e.ts
@@ -1,42 +1,41 @@
 import { expect } from '@playwright/test';
 import { configs, test } from '@utils/test/playwright';
 
+import { RadioFixture } from '../fixtures';
+
 /**
  * This behavior does not var across modes/directions.
  */
 configs({ modes: ['ios'], directions: ['ltr'] }).forEach(({ title, config }) => {
-  test.describe(title('radio-group'), () => {
-    test.beforeEach(async ({ page }) => {
+  test.describe(title('radio-group: state'), () => {
+    test('radio should remain checked after being removed/readded to the dom', async ({ page }) => {
       await page.goto('/src/components/radio-group/test/search', config);
-    });
 
-    test.describe('radio-group: state', () => {
-      test('radio should remain checked after being removed/readded to the dom', async ({ page }) => {
-        const radioGroup = page.locator('ion-radio-group');
-        const radio = page.locator('ion-radio[value=two]');
-        const searchbarInput = page.locator('ion-searchbar input');
+      const radioFixture = new RadioFixture(page);
 
-        // select radio
-        await radio.click();
-        await expect(radio.locator('input')).toHaveJSProperty('checked', true);
+      const radioGroup = page.locator('ion-radio-group');
+      const searchbarInput = page.locator('ion-searchbar input');
 
-        // filter radio so it is not in DOM
-        await page.fill('ion-searchbar input', 'zero');
-        await searchbarInput.evaluate((el) => el.blur());
-        await page.waitForChanges();
-        await expect(radio).toBeHidden();
+      // select radio
+      const radio = await radioFixture.checkRadio('mouse', 'ion-radio[value=two]');
+      await radioFixture.expectChecked(true);
 
-        // ensure radio group has the same value
-        await expect(radioGroup).toHaveJSProperty('value', 'two');
+      // filter radio so it is not in DOM
+      await page.fill('ion-searchbar input', 'zero');
+      await searchbarInput.evaluate((el) => el.blur());
+      await page.waitForChanges();
+      await expect(radio).toBeHidden();
 
-        // clear the search so the radio appears
-        await page.fill('ion-searchbar input', '');
-        await searchbarInput.evaluate((el) => el.blur());
-        await page.waitForChanges();
+      // ensure radio group has the same value
+      await expect(radioGroup).toHaveJSProperty('value', 'two');
 
-        // ensure that the new radio instance is still checked
-        await expect(radio.locator('input')).toHaveJSProperty('checked', true);
-      });
+      // clear the search so the radio appears
+      await page.fill('ion-searchbar input', '');
+      await searchbarInput.evaluate((el) => el.blur());
+      await page.waitForChanges();
+
+      // ensure that the new radio instance is still checked
+      await radioFixture.expectChecked(true);
     });
   });
 });
